Initial commit
The xdg-mime-rs is a reimplementation in Rust of the xdgmime C library,
which is used to read and query the shared-mime-info database of
associations between MIME types and file types. The shared-mime-info
database is typically used on all modern Linux and Unix-like operating
systems as a way to determine the file type from both the extension and
"magic" markers inside the file.

For more information see:

 - https://gitlab.freedesktop.org/xdg/xdgmime
 - https://gitlab.freedesktop.org/xdg/shared-mime-info