glob: pass glob strings as &str, not Into<String> generics
This minimizes string copies, and for callers it's not a problem to do

  let foo = String::from("foo*");

  let glob = Glob::simple(mime_type, &foo);

i.e. pass it as "&foo", since String derefs to &str anyway with the &
operator.

The alternative is to use AsRef<str> in public APIs and then convert
to &str immediately, to avoid all the internals getting monomorphized
everywhere.