alias: separate ordering and equivalence concerns
Similar to the last commit, we use sort_by() and remove the funky
implementations of Ord/PartialOrd for Alias.

We #[derive(PartialEq)] on Alias to have its usual behavior, in this
case, letting us compare Alias objects in tests with assert_eq!().

We add a new Alias::is_equivalent() method, which has the old
PartialEq behavior, and update the corresponding test to use it.