Add API to guess a MIME type
We provide the low level API to determine a MIME type for a given file
or chunk of data, but it'd be more convenient for everyone if we had a
method to guess the MIME type with the appropriate algorithm defined by
the shared-mime-info specification.

Since you can specify either, or both, the file name and the chunk of
data, the most idiomatic approach is to have a builder pattern:

https://doc.rust-lang.org/1.0.0/style/ownership/builders.html

We ask SharedMimeInfo to give us an object to fill out, tied to the
lifetime of the SharedMimeInfo instance, and ask the object to guess.

In the future we can have additional fields on the builder type, like
the full path of the file name, so that it can read its contents
automatically; or a std::fs::Metadata instance, so we can check if the
file is zero-sized.