magic: Fix the rule matching
The rule traversal is incorrect when it comes to matching rules that are
at the same level. We're only bailing out of the loop if this is the
last rule, but we have a negative match for rules that match in the
middle of an entry.

This is easy to verify with shell scripts; the match entry for
application/x-shellscript has a lot of rules with the same indent level,
but we only ever match data that satisfies the last rule.

Fixes: #13
(cherry picked from commit 1a57dd237bc9c0306e17cee0389813931d17026a)
Signed-off-by: Emmanuele Bassi <ebassi@gnome.org>